Refactored constitutive_laws
Refactored constitutive laws created a sub-module. This will permit in the future (or even now) to split constitutive laws in separate files making it more easy to be maintained.
